Corrects in HQService

+++ src/main/java/com/hospitalwardmanagement/service/implementation/HealthQuestionnaireServiceImplementation.java
@@ -1,6 +1,7 @@
 package com.hospitalwardmanagement.service.implementation;
 
 import com.hospitalwardmanagement.exceptions.HealthQuestionnaireExistsException;
+import com.hospitalwardmanagement.exceptions.HealthQuestionnaireNotExistForPatientException;
 import com.hospitalwardmanagement.exceptions.ResourceNotFoundException;
 import com.hospitalwardmanagement.model.healthQuestionnaire.HealthQuestionnaire;
 import com.hospitalwardmanagement.model.patient.Patient;
@@ -40,17 +41,28 @@ public class HealthQuestionnaireServiceImplementation implements HealthQuestionn
     @Override
     public HealthQuestionnaire updateHealthQuestionnaireByPatientId(HealthQuestionnaire healthQuestionnaire, Long patientId) {
         Patient patient = getPatient(patientId);
+
+        if(patient.getHealthQuestionnaire()==null)
+            throw new HealthQuestionnaireNotExistForPatientException(patientId);
+
         Long existingHealthQuestionnaireId = patient.getHealthQuestionnaire().getId();
-        HealthQuestionnaire existingHealthQuestionnaire = getHealthQuestionnaireById(existingHealthQuestionnaireId); // sprawdzić jak sie zachowa gdy dostaje null
+        HealthQuestionnaire existingHealthQuestionnaire = getHealthQuestionnaireById(existingHealthQuestionnaireId);
+
         existingHealthQuestionnaire.setHeight(healthQuestionnaire.getHeight() == null ? existingHealthQuestionnaire.getHeight() : healthQuestionnaire.getHeight());
         existingHealthQuestionnaire.setWeight(healthQuestionnaire.getWeight() == null ? existingHealthQuestionnaire.getWeight() : healthQuestionnaire.getWeight());
         existingHealthQuestionnaire.setMedicalHistories(healthQuestionnaire.getMedicalHistories() == null ? existingHealthQuestionnaire.getMedicalHistories() : healthQuestionnaire.getMedicalHistories());
-        existingHealthQuestionnaire.setPatient(healthQuestionnaire.getPatient() == null ? existingHealthQuestionnaire.getPatient() : healthQuestionnaire.getPatient());
+
         return questionnaireRepository.save(existingHealthQuestionnaire);
     }
 
     @Override
     public void deleteHealthQuestionnaireByPatientId(Long patientId) {
+        Patient patient = getPatient(patientId);
+        if(patient.getHealthQuestionnaire()==null)
+            throw new HealthQuestionnaireNotExistForPatientException(patientId);
+
+        patient.setHealthQuestionnaire(null);
+
         HealthQuestionnaire healthQuestionnaireToDelete = getHealthQuestionnaireByPatientId(patientId);
         questionnaireRepository.delete(healthQuestionnaireToDelete);
     }
@@ -59,18 +71,15 @@ public class HealthQuestionnaireServiceImplementation implements HealthQuestionn
     @Override
     public HealthQuestionnaire getHealthQuestionnaireByPatientId(Long patientId) {
         Patient patient = getPatient(patientId);
+        if(patient.getHealthQuestionnaire()==null)
+            throw new HealthQuestionnaireNotExistForPatientException(patientId);
+
         Long existingHealthQuestionnaireId = patient.getHealthQuestionnaire().getId();
-        HealthQuestionnaire healthQuestionnaire = getHealthQuestionnaireById(existingHealthQuestionnaireId);
-        return healthQuestionnaire;
+        return getHealthQuestionnaireById(existingHealthQuestionnaireId);
     }
 
     public HealthQuestionnaire getHealthQuestionnaireById(Long id) {
-        if (id == null)
-            throw new NullPointerException();
-
-        return questionnaireRepository.findById(id).orElseThrow(
-                () -> new ResourceNotFoundException("Health Questionnaire", "id", id)
-        );
+        return questionnaireRepository.findById(id).get();
     }
 
 
@@ -78,6 +87,6 @@ public class HealthQuestionnaireServiceImplementation implements HealthQuestionn
         return patientRepository.findById(patientId).orElseThrow(
                 () -> new ResourceNotFoundException("Patient", "id", patientId)
         );
-
     }
+
 }
